Codegen: Outsource the linear algebra c functions (inverse, solve) into separate files and read them in as strings for the preamble of the respective loopy Callable.
